+++ src/Starcounter.XSON/Class Json/Json.State.cs
@@ -1,48 +1,65 @@
-﻿using Starcounter.Advanced;
-using Starcounter.Internal.XSON;
-using Starcounter.Templates;
-using System;
+﻿using System;
 using System.Collections;
 using System.Collections.Generic;
 using Starcounter.Internal;
+using Starcounter.Internal.XSON;
+using Starcounter.Logging;
+using Starcounter.Templates;
 using Starcounter.XSON;
 
 namespace Starcounter {
     partial class Json {
+        [Flags]
+        protected enum PropertyState : byte {
+            Default = 0,
+            /// <summary>
+            /// The property is considered dirty and value should be sent to client.
+            /// </summary>
+            Dirty = 1,
+            /// <summary>
+            /// Set for bound properties. The bound value is cached and should be read from
+            /// the cache. Used to avoid changes in bound values while gathering changes and 
+            /// serializing since these parts are done in steps.
+            /// </summary>
+            Cached = 2
+        }
+
+        private static LogSource logSource = new LogSource("Starcounter.XSON");
 
         /// <summary>
         /// Backing field for the transaction applied to this instance (if any).
         /// </summary>
-//        private ITransaction _transaction;
-        private TransactionHandle _transaction;
+        private TransactionHandle transaction;
 
         /// <summary>
         /// Cache element index if the parent of this Obj is an array (Arr).
         /// </summary>
-        internal int _cacheIndexInArr;
+        internal int cacheIndexInArr;
 
         /// <summary>
         /// Tells if any property value has changed on this container (if it is an object) or
         /// any of its children or grandchildren (recursivly). If this flag is true, there can be
         /// no changes to the JSON tree (but there can be changes to bound data objects).
         /// </summary>
-        internal bool _Dirty = false;
-
-        //private bool __BrandNew_ = true;
-
+        internal bool dirty = false;
+        
         /// <summary>
         /// 
         /// </summary>
-        internal bool _PendingEnumeration;
+        internal bool pendingEnumeration;
 
         /// <summary>
-        /// For unbound values, we keep a list of flags to know which properties has changed.
+        /// A list containing state for each property. Will only be initialized if changes are tracked.
         /// </summary>
-        protected List<bool> _SetFlag;
+        protected List<PropertyState> stateFlags;
 
         /// <summary>
+        /// If this is a dynamic or a non-codegenerated jsonobject the list will contain the values 
+        /// for each property.
+        /// If this is an array, the list will contain the rows.
+        /// If none of the above the list is not used and will never be initialized.
         /// </summary>
-        private IList _list;
+        private IList valueList;
 
         /// <summary>
         /// Json instances (objects or arrays) can be values in a hosting object property or in a 
@@ -51,17 +68,17 @@ namespace Starcounter {
         /// have a single parent. Our implementation provides the service of finding the declaring
         /// (parent) object of this object.
         /// </summary>
-        private Json _parent;
+        private Json parent;
 
         /// <summary>
         /// Json objects can be stored on the server between requests as session data.
         /// </summary>
-        internal Session _Session;
+        internal Session session;
 
         /// <summary>
         /// Keeps track on when we added/inserted or removed elements
         /// </summary>
-        internal List<Change> ArrayAddsAndDeletes = null;
+        internal List<Change> arrayAddsAndDeletes = null;
 
         /// <summary>
         /// 
@@ -76,12 +93,12 @@ namespace Starcounter {
         /// <summary>
         /// Implementation field used to cache the Metadata property.
         /// </summary>
-        private ObjMetadata<TObject, Json> _Metadata = null;
+        private ObjMetadata<TObject, Json> metadata = null;
 
         /// <summary>
-        /// The _ template
+        /// The template this object is based on.
         /// </summary>
-        internal Template _Template;
+        private Template template;
 
         /// <summary>
         /// An Json object or array can be bound to a data object. This makes the Json reflect the data in the
@@ -89,29 +106,29 @@ namespace Starcounter {
         /// or view models are often associated with database objects. I.e. a person form might
         /// reflect a person database object.
         /// </summary>
-        internal object _data;
+        internal object data;
 
         /// <summary>
-        /// List containing all stepsiblings that exists on this level.
+        /// List containing all siblings that exists on this level.
         /// </summary>
-        private SiblingList _stepSiblings;
+        private SiblingList siblings;
 
-        internal String _appName;
+        internal String appName;
 
-        internal Boolean _wrapInAppName;
+        internal Boolean wrapInAppName;
         
         /// <summary>
         /// If set to true, additional features for keeping track of changes and getting a log of changes 
         /// are initialized. If not needed this should not be enabled since the performance will be much worse.
         /// </summary>
-        private bool _trackChanges;
+        private bool trackChanges;
 
         /// <summary>
         /// If set to false, bound properties will not be updated automatically.
         /// </summary>
-        internal bool _checkBoundProperties;
+        internal bool checkBoundProperties;
  
-        private bool _isArray;
+        private bool isArray;
 
         /// <summary>
         /// If this json is a part of a stateful viewmodel (i.e. puppet) this field contains