+++ src/Starcounter.XSON/Sessions/Session.cs
@@ -81,6 +81,12 @@ namespace Starcounter {
         /// </summary>
         internal bool enableNamespaces = false;
 
+        /// <summary>
+        /// If set to true values that are bound will not be read several times when generating
+        /// changes from changelog and creating patches.
+        /// </summary>
+        internal bool enableCachedReads = false;
+
         public Session() : this(SessionOptions.Default) {
         }
 
@@ -370,7 +376,7 @@ namespace Starcounter {
                     if (value.Parent != null)
                         throw ErrorCode.ToException(Error.SCERRSESSIONJSONNOTROOT);
 
-                    if (value._Session != null && value._Session != this)
+                    if (value.session != null && value.session != this)
                         throw ErrorCode.ToException(Error.SCERRJSONSETONOTHERSESSION);
                 }
 
@@ -386,7 +392,7 @@ namespace Starcounter {
                     }
 
                     if (value != null) {
-                        value._Session = this;
+                        value.session = this;
 
                         if (publicViewModelIndex == -1)
                             publicViewModelIndex = stateIndex;
@@ -560,12 +566,12 @@ namespace Starcounter {
                 return true;
             }
         }
-
+        
         /// <summary>
         /// 
         /// </summary>
         public bool HaveAddedOrRemovedObjects { get; set; }
-
+        
         internal TransactionHandle RegisterTransaction(TransactionHandle handle) {
             TransactionRef tref = null;