+++ src/Starcounter/Apps/JsonResponseMerger.cs
@@ -59,16 +59,16 @@ namespace Starcounter.Internal {
                 mainJson = resp.Resource as Json;
 
                 if (mainJson != null) {
-                    if (mainJson.StepSiblings != null) {
-                        if (mainJson.StepSiblings.HasBeenSent(mainJson.StepSiblings.IndexOf(mainJson))) {
+                    if (mainJson.Siblings != null) {
+                        if (mainJson.Siblings.HasBeenSent(mainJson.Siblings.IndexOf(mainJson))) {
                             stepSiblings = new SiblingList();
                             stepSiblings.Add(mainJson);
                             stepSiblings.MarkAsSent(0);
-                            mainJson.StepSiblings = stepSiblings;
+                            mainJson.Siblings = stepSiblings;
                         }
                     }
-                    mainJson._appName = resp.AppName;
-                    mainJson._wrapInAppName = true;
+                    mainJson.appName = resp.AppName;
+                    mainJson.wrapInAppName = true;
                 }
 
                 return resp;
@@ -102,14 +102,14 @@ namespace Starcounter.Internal {
 
             if (mainJson != null) {
 
-                mainJson._appName = mainResponse.AppName;
-                mainJson._wrapInAppName = true;
+                mainJson.appName = mainResponse.AppName;
+                mainJson.wrapInAppName = true;
                 
-                var oldSiblings = mainJson.StepSiblings;
+                var oldSiblings = mainJson.Siblings;
 
                 stepSiblings = new SiblingList();
                 stepSiblings.Add(mainJson);
-                mainJson.StepSiblings = stepSiblings;
+                mainJson.Siblings = stepSiblings;
 
                 if (responses.Count == 1) {
                     MarkExistingSiblingsAsSent(mainJson, oldSiblings);
@@ -133,24 +133,24 @@ namespace Starcounter.Internal {
                         if (siblingJson == null)
                             continue;
 
-                        siblingJson._appName = responses[i].AppName;
-                        siblingJson._wrapInAppName = true;
+                        siblingJson.appName = responses[i].AppName;
+                        siblingJson.wrapInAppName = true;
 
-                        if (siblingJson.StepSiblings != null) {
+                        if (siblingJson.Siblings != null) {
                             // We have another set of step-siblings. Merge them into one list.
-                            foreach (var existingSibling in siblingJson.StepSiblings) {
+                            foreach (var existingSibling in siblingJson.Siblings) {
                                 // TODO:
                                 // Filtering out existing siblings that comes from the same app.
                                 // This is a hack to avoid having multiple layouts from the launcher
                                 // that gets merged, since the merger gets called a lot.
                                 // Proper solution needs to be investigated.
                                 // Issue: https://github.com/Starcounter/Starcounter/issues/3470
-                                if (stepSiblings.ExistsForApp(existingSibling._appName))
+                                if (stepSiblings.ExistsForApp(existingSibling.appName))
                                     continue;
                                 
                                 if (!stepSiblings.Contains(existingSibling)) {
                                     stepSiblings.Add(existingSibling);
-                                    existingSibling.StepSiblings = stepSiblings;
+                                    existingSibling.Siblings = stepSiblings;
                                 }
                             }
                         }
@@ -158,7 +158,7 @@ namespace Starcounter.Internal {
                         if (!stepSiblings.Contains(siblingJson)) {
                             stepSiblings.Add(siblingJson);
                         }
-                        siblingJson.StepSiblings = stepSiblings;
+                        siblingJson.Siblings = stepSiblings;
                     }
                 }
 
@@ -177,7 +177,7 @@ namespace Starcounter.Internal {
             if (json == null || afterMergeCallbacks_.Count == 0)
                 return;
 
-            list = json.StepSiblings;
+            list = json.Siblings;
             if (list == null) {
                 list = new SiblingList();
                 list.Add(json);
@@ -186,14 +186,14 @@ namespace Starcounter.Internal {
             foreach (var hook in afterMergeCallbacks_) {
                 newSibling = hook.Invoke(request, callingAppName, list);
                 if (newSibling != null) {
-                    newSibling._wrapInAppName = true;
+                    newSibling.wrapInAppName = true;
                     list.Add(newSibling);
-                    newSibling.StepSiblings = list;
+                    newSibling.Siblings = list;
                 }
             }
 
-            if (json.StepSiblings == null && list.Count > 1) {
-                json.StepSiblings = list;
+            if (json.Siblings == null && list.Count > 1) {
+                json.Siblings = list;
             }
         }
 
@@ -201,7 +201,7 @@ namespace Starcounter.Internal {
             SiblingList newSiblings;
 
             if (oldSiblings != null && mainJson.Parent != null) {
-                newSiblings = mainJson.StepSiblings;
+                newSiblings = mainJson.Siblings;
                 for (int i = 0; i < newSiblings.Count; i++) {
                     int index = oldSiblings.IndexOf(newSiblings[i]);
                     if (index != -1) {